Refactoring RPC components
This is a work-in-progress state of affairs, namely it changes the way
that RPC core components (aimed at qaul developers) and SDK
components (aimed at application developers) are identified.  This
commit also breaks user storage (search for `todo!()`) which means
that this state needs to be fixed before being usable again.