docs/website: allow HTML markup in Markdown files
This option is unhelpfully called "unsafe".  The reason for this (and
for it not being on by default), is that Hugo's BDFL is worried that
with "content contributed from millions of people on the interweb (via
PRs and similar), it's pretty easy to envision that that editor will
let some unsafe content slip through and deployed".  I don't think
that's really relevant to us.

Allowing HTML markup means we can use semantic HTML like the <address>
element, for which there is no Markdown syntax.

[1]: https://github.com/gohugoio/hugo/issues/6581#issuecomment-562102816