docs/website: fix only child navbar links alignment
.navbar-links is designed to always be aligned to the right.
Currently, this is achieved by putting it as the second element inside
a flex container with justify-content: space-between, which pushes it
over to the right.  But that doesn't work when it's the only item in a
flex container, because the flex items still start at the left.  To
resolve that case, we can set margin-left: auto.  This has no effect
in the multiple-element case, but solves the problem when
.navbar-links is an only child.