docs/website: use semantic structure elements
HTML has dedicated elements for lots of these kinds of things, so we
can improve readability and semantic meaning by using those instead of
divs.

I changed the .header and .footer styles to body > header and body >
footer, because I don't think it's worth having a special class that
would just be identical to the element names, but we might want to
have inner headers and footers at some point, so I think it's worth
restricting to just ones that are direct body descendants.