libqaul: fix tests
This commit changes the way that libqaul handles UserUpdates, in
accordance with the recent change in libqaul_types.

Some of the integration tests started aborting and I'm not sure why.
I set them all to 'ignore' for now so that we can at least have some
of the tests.  This will require further investigation.