docs/website: pin footer to bottom of viewport
Previously, the footer would show up immediately after the content on
short pages, so it would appear in the middle of the page, with blank
space after it.  We can solve that by making all blocks (apart from
the header and footer) grow equally until they've filled up the page,
making sure the footer is no higher than the bottom of the window.

We have to set align-items center, or otherwise the content will
appear at the top of a stretched block, which looks fine on a page
with just a single stretched block, but not so good on a page with
several like the home page.