Replace failure with thiserror
failure is unmaintained.  The replacement for the derive bits of
failure (which is all we were using) is thiserror, so switch to using
that.

The main difference is that failure had to implement a bunch of stuff
itself that has since been added to the standard library, so the
replacements use the stuff in std rather than rolling their own.