deny.toml: create
This is a configuration file for cargo-deny, which can do licensing
and security checks for us.  I've set it up to check for security
advisories, license compliance, and unrecognized git sources (no
recognized git sources are currently configured).

License compliance for qaul is a bit complicated because of our
unusual licensing, so I've included some guidance for adding a
dependency with a license that isn't already allowlisted that is
hopefully easy enough to follow.