irdest-core: improve the service storage API and store internals
Previously each service had access to many key-value stores, exposed
via the `MetadataMap` type.  The problem with this approach was that
services were responsible for differentiating between insertions and
updates, because this is a distinction that Alexandria enforces.

This API now exposes a single key-value store per service, with three
function endpoints as documented in #7: insert, delete, and query.
Insert acts as an upsert variant, which will either insert or update a
single key in the key-value store associated to the service.

In order to allow services to better namespace their data the key type
is `StoreKey`, a 2-string tuple which provides a namespace, alongside
a data key