alexandria: swap capnproto for protobuf crate
While we were using Capnproto for the RPC protocol before and it has
some advantages over protobuf, the Rust crate is just absolutely
unusable trash.  Writing this code was basically impossible.  And
maybe the library wants to stop you from doing it, but that's the only
thing I want to use the library _for_.  We should however definitely
investigate upgrading to protobuf 3.0 when it's ready.