ratman: change the way local addresses are treated
Previously we could rely on irdest-core to handle the local address
case for us where the dispatch logic in Ratman didn't have to
understand how to send messages to itself.  But since the architecture
has now changed to one where many clients are connected directly to a
Ratman daemon, this is a required feature.

This change is a little bit dirty, i.e. it doesn't distinguish between
"local address" and "address doesn't exist".  In reality the daemon
layer will not forward a message to an address that doesn't exist, but
this will create a lot of error logging and should be fixed in the future!