Cleanup stale project files
This deletes _a lot_ of code, some of which we may want to recover in
the future.  We should get the code that was deleted from the history
of the repository, instead of carrying around countless crates.  This
is done to limit the scope of the project and make it easier to think
about the ways we may want to refactor the codebase in the future.