netmod-inet: fix peering issues (#35)
Two things were at fault:

1. We removed the keep-alive part of the HELLO handshake.  Ratman
already frequently sends data via a link (Address announcements) so
this is most likely not needed.  Furthermore it made  it hard to
properly handle because now HELLO packets were context sensitive.

2. When establishing a connection to a peer we included the WRONG PORT
in the HELLO packet, which likely was not caught until now because we
were always running on the same port (but different computers).

This commit may not fix all issues with netmod-inet but it _should_
make it more stable and resilient.

In the future we may want to look into re-writing parts of this driver
to be easier to handle.  Especially using a split I/O stream to avoid
having to mutex access.