ratman: embed and serve webui
This embeds the built Ember.js Web UI into the binary (release), or reads
them from the source tree (development), which means we don't have to worry
about where on disk the assets are relative to the binary.

But it comes with the caveat that building `ratmand` with `--feature=webui`
will now error unless you've previously ran `cd ratman/webui; yarn build`.