ratman/dashboard: use browser language where possible
`ember-best-language` is an ember.js helper that tries to find a match
between the application's available languages, and `navigator.languages`.

Currently, we support English (en) and German (de), and the addon correctly
handles regional fallbacks, eg. US English (en-US) and Austrian (de-AT).

If we get eg. a user who speak only Swiss Italian and Japanese, we log a
warning (to debug faulty locale resolution), and fall back to English:

  [WARN] 🏳 No localisation for [ "it-CH", "ja" ] only [ "de", "en" ]