ratman: ratman: fixing tcp tests with sleep
This is an interesting race-condition where the peering between the
inet instances hasn't completed yet before the test sends a flood
message.  The message is thus not sent to anyone because the router
can't even know that the peer was offline.

In a real setting this would be very unlikely to happen.

However, there are some things we can learn from this:

1. for outgoing peering sessions we should add an API to netmod to
allow the router to query the peering state to indicate whether a
frame should be included in the journal if it is a flood message

2. the journal will fix many of these issues

3. some mechanism for the router to indicate that it has successfully
peered that we can rely on in CI for more complicated test setups