ratman: refactor dependency structure
This commit touches a lot of code and simplifies some things.
Previously we had three different abstractions for Recipients, now we
only use one (or two, depending on how you're counting).

This also moves Frame, Recipient, and the sequence builder utilities
to ratman-types and adding this as a dependency to ratman-netmod.  It
also removes the ratman-identity library and foldes this type into
ratman-types.

This also changes the paths of the libraries to make it more
consistent.