ratman-types: fix protobuf syntax to remove illegal statements
In the proto3 syntax the "optional" keyword is apparently no longer
allowed.  There is some inconsistent behaviour in the protobuf tooling
around this, since there are versions of protoc with which this
compiles fine, and others which throw an error.

The other side-effect is that the generated Rust code is different
whether an `optional` statement is included or not.  Previously we
included them because it turned `optional bytes` into
`Option<Vec<u8>>`, which is a very sensible way to encode this data.
Without the `optional` statements it generates `Vec<u8>` which we then
have to check for whether it is empty or not.  Same
behaviour (hopefully), but more annoying.

We will want to do some research into this and why there are versions
of protoc that allow this and others that don't...