ratman: restructure setup trampoline
In the future we will need to move even more code so that integration
of ratman into a library can be easier.

The question is whether we should map the Ratman API or just allow
linking locally, while still using the IPC socket for communicating
with the ratman daemon machinery.  The former will be cleaner, the
latter faster and easier to build.