ratman: clean-up error system
Removed deprecated error handling module and unified all error
handling under ratman_types::Error.

This system is not perfect and will need to be cleaned up in the
future.  But maybe it will be easier to do once the new journal and
collector systems are implemented and we have a better grasp at what
error types have to exists.