ratman/dashboard: Refresh models periodically
Having auto-updating metrics on the front page is imo neat, and also kind
of necessary because of how the metrics work (deltas between polls).
But if one thing is visibly auto-updating, and nothing else does (new peers
need a full-page load to show up), that's not neat, it's just confusing UX.

This polls the model of all active routes for updates every 5s, waiting for
the previous refresh to finish before kicking off another one to avoid a
thundering herd.

---------------------------------------------------------------------------

In webapp terms, a poll every 5s might sound like a lot, but remember that
we're expecting this to have roughly one concurrent user, on the same host
as the daemon, making 2 requests (metrics + model) every 5s.

If that starts causing serious load on the daemon, that's a bug in the API.