ratman: move lib-netmod, lib-client, and lib-types into libratman
The types library was previously required by the ratman daemon,
client-lib, and netmod-lib.  While it doesn't make much sense for an
application to bind both against the application and netmod APIs,
consolidating this code makes it easier to manage dependencies.

Compile-time features could also be used to enable and disable
features to make the final compile target smaller.