ratman: library / binary split refactoring
See the previous commit messages for additional rationale on this
change.

* Move most of the ratman codebase back into the compiled scope and
update references to the previously split ratman libraries.

* Update existing crate dependencies and remove some that are no
longer needed, or will only be needed later on in development

* Update the x25519-dalek crate to the latest 2.0.0 release candidate

* Rename the router crate from "ratman" to "ratmand" to reflect its
nature as a daemon that is either run via the "ratmand" binary target
or can be embedded as a library into other crates (ffi interface
missing)

* Rename "libratman" to "ratman" to replace this crate on crates.io.
In the majority of cases someone looking to interact with an Irdest
network will want to write a ratman client or network driver.  Both
tasks are now handled by a single library, which removes the need to
specify the type of library.  Furthermore having the daemon code on
crates.io under such a prominent name makes little to no sense.

This means that the following crates must now be deprecated:

* ratman-client
* ratman-netmod
* ratman-identity
* ratman-types
* ratman-harness
* ratman-configure