Various Ratman updates
- Create a better abstraction over active client connections coupled
to existing clients in the registry.  This still needs some work, and
generally it has become obvious that all this data needs to be backed
by the database already.  So that's the next step in the process.

- Moving the configuration module and including the KDL crate to start
porting over the configuration.  More notes on this will follow, but
hopefully by the end of this the configuration is much easier to read
and write, and self-documenting.

- Moving over little bits and pieces from the "old" tree and removing
them from those respective files