Refactor error facades and netmod initialisation
This is a rather large commit which re-introduces some more code into
the compilation scope.  Initialisation of netmods has been moved to
'config/startup', and base-initialisation and peer initialisation has
been split.  This allows (at least in theory) peers to be added to a
driver after runtime has already started.

We new also compile the lora netmod again, and this commit fixes
relevant issues that came up from the libratman consolidation/
refactor.  All in all, we're getting closer to being able to start the
router back up.  I assume many other small things will break.  I've
also not been able to test the configuration parsing yet, but from my
understanding of how the kdl API works, this usable _should_ be valid.

One thing to keep an eye on is that we now have two almost identical
abstractions to express a "driver map".  One is pre-startup, where
keys are strings (the name of the netmod from the configuration), and
one internal to the routing core, where keys are numeric IDs.  Maybe
we can get rid of one, or we may want to rename one to be less confusing.