Fix test compilation issues and expose a way to override test config
The `patch` function is very hacky and should never actually be used
in production (which is why it's cfg(test) to represent that fact).
But it'll be useful to be able to quickly override the default
configuration for tests (annoyingly apparently "override" is a
reserved keyword ??).

This commit also starts working on properly integrating the routing
core.  Previously we had a (hopefully) correctly initialised list of
netmods/ drivers, now we add the peers to them, and then add each
endpoint to the routing core.  I simplified the API for the routing
core to no longer rely on anonymous generics (or any generics), since
the only consumer of this API is now the daemon/ router context.  This
means we can just juggle trait objects from start to finish and things
are good and easy.