Add a set of carrier frame prototypes and update the specification
This commit adds three prototypes on how to encode the basic frame
formats (and higher level structures too).  The three encoding systems
on trial here are:

bincode - written for the servo engine, which produces a very fast,
albeit relatively big binary format.

protobuf - the same encoding schema we're already using in the client
API, which has language support in various languages, and creates a
lot of abstractions over data, versioning, etc

hand-rolled - this approach takes the current specification, and
applies it in a very binary-literal way.  The resulting format is the
same size as the protobuf generated code (at least in the announcement
use-case), while being nicer to read in a hexdump.

This code also adds a full carrier frame format and some encoding
utilities.

Importantly, this adds a lot of overhead to the generating and parsing
of binary frames.  The next version of this code should be ported to
nom [1] to make adding new frame types less code-heavy.

[1]: https://docs.rs/nom/