ratmand: implement new carrier frame slicer on eris block input
Some changes to the underlying carrier frame format were needed.
Previously the sequence Id was a simple 32-byte ID.  This was great to
align the block reference ID, but didn't leave any room to also keep
track of frame order.  Since frames MAY arrive out of order we need a
way to re-associate them.

A previous system for re-ordering relied on nested hashes for content
validation.  Since the top-level integrity is validated via the Eris
content hash (and the signed block manifest) anyway, we can dispense
with this, and replace this mechanism with a simple u8 counter.

Since each block sequence still uses a different block reference index
this is only to allow for frames per block.  It's unlikely that a
single ERIS block would result in more than 255 carrier frames being
generated, even on very small MTUs.