netmod-inet: update to new carrier frame format
This module still uses its own session error type which is a bit too
annoying to disentangle from the core of it right now.  We'll leave it
as is now.  Since we still need to update other errors around the
Irdest stack, it's probably not too much of a big deal.