ratmand: remove async_std environment from program root
This changes the way that async components are called.  Instead of
handing of the entire main function to async_std, this commit makes
each invocation of async components explict.  This separates the
memory contexts between the allocation phase and the runtime phase of
the router, hopefully resolving stack pressure from more data
intensive interactions.