ratmand: fix journal block collection mechanism
This is a pretty big commit, but none of the changes make sense
without everything else in here (ok there are some trace logging
changes, to make the log less spammy and slightly more useful).

Mostly this adds some new tasks for re-assembling blocks back into
messages and making the message sending mechanism use the new block
and frame slicer, sending the manifest, etc.

There's considerable delay between collecting blocks and receiving the
manifest, and so currentnly messages have about a 1 second delay where
the router waits until it tries again to re-assemble a message from
blocks.  This should be tweaked, and maybe we can also reduce the
amount of latency in the collector.

Overall, this touches a lot of things, but makes the whole stack work
again!