Fix deadlock issues by making async systems multithread by default
Currently the announcer system runs on the same task thread as the API handler,
which means that a blocking receive or a long send will halt announcements.  The
real solution is to have a dedicated announcement system and connect them with
channels.  But I'm lazy right now and I wanna get this release done so let's
just move that to the next one